generate snapshot
