+++ pkg/tests/apis/openapi_snapshots/dashboard.grafana.app-v0alpha1.json
