+++ pkg/tests/apis/openapi_snapshots/folder.grafana.app-v0alpha1.json
