+++ pkg/tests/apis/openapi_snapshots/peakq.grafana.app-v0alpha1.json
